+++ packages/cli/CHANGELOG.md
@@ -3,12 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [0.27.13](https://github.com/zetapush/zetapush/compare/v0.27.12...v0.27.13) (2018-06-26)
+# [0.28.0](https://github.com/zetapush/zetapush/compare/v0.27.14...v0.28.0) (2018-06-29)
+
+
+### Bug Fixes
+
+* **cli:** replace ascii-progress by node-progress-bars ([fdfdca2](https://github.com/zetapush/zetapush/commit/fdfdca2)), closes [#55](https://github.com/zetapush/zetapush/issues/55)
+
+
+### Features
+
+* **cli:** according to documentation, rename http-server option to serve-front ([7a0268d](https://github.com/zetapush/zetapush/commit/7a0268d)), closes [#55](https://github.com/zetapush/zetapush/issues/55)
+
+
+### BREAKING CHANGES
+
+* **cli:** rename http-server option to serve-front
 
 
 
-
-**Note:** Version bump only for package @zetapush/cli
 
 <a name="0.24.1"></a>
 ## [0.24.1](https://github.com/zetapush/zetapush/compare/v0.24.0...v0.24.1) (2018-06-18)