+++ packages/platform/CHANGELOG.md
@@ -3,12 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [0.27.13](https://github.com/zetapush/zetapush/compare/v0.27.12...v0.27.13) (2018-06-26)
+# [0.28.0](https://github.com/zetapush/zetapush/compare/v0.27.14...v0.28.0) (2018-06-29)
 
 
 ### Features
 
-* **platform:** improve typings for auth platforms ([c06ac28](https://github.com/zetapush/zetapush/commit/c06ac28))
+* **platform:** add typings for Gda service ([dd64809](https://github.com/zetapush/zetapush/commit/dd64809))